web: behaviour: instance groups in sidebar
concourse/concourse#5921

Driving the tooltip behaviour with tests wasn't obvious.

We'll eventually want to pull the badge view into a Views helper
eventually (perhaps when we add the breadcrumbs that also include the
badge).

We'll also want to extract the logic of computing instance groups
somewhere - though, I'm not sure exactly where this belongs.

One thing that I wasn't sure about is how we want to display the
following case in the favorites section:
* There's an instance group that contains a pipeline that has no
  instance vars
* Only that pipeline is favorited
This applies to the dashboard as well - do we want to display a regular
pipeline card/list item, or an instance group with only 1 element?

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>
Co-authored-by: Bohan Chen <bochen@pivotal.io>