web: structure: add instance group id to dashboard route
concourse/concourse#5921

I debated whether or not to make the instance group page a completely
separate view from the dashboard. I decided that it would be better to
keep it on the dashboard - to indicate that you're viewing an instance
group, you can use the `?team=...&group=...` query parameters.

I had discussed with @matthewpereira an approach where the instance
group view is just a filtered down dashboard - clicking on the instance
group name is a shorthand for applying a filter to view only the
specific instance group. While this requires more work to get right
(esp. the ability to apply multiple filters simultaneously), modeling
"viewing an instance group" as an extension of the `SearchType` on the
dashboard feels like a reasonable compromise, and should make it
relatively easy to transition over to a native dashboard filter if we
choose to do so.

Note that this is marked as structure as we're currently just ignoring
the selected instance group, if there is one.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>