web: behaviour: instance group links to instance group view
concourse/concourse#5921

...in both the dashboard and sidebar.

This would be more interesting if the instance group view was at all
different from the dashboard, but one step at a time!

One thing that's missing is that the sidebar does not reflect the fact
that we're viewing an instance group by having a dark background. This
is a bit tricky, as the sidebar needs to know what route we're currently
on. The sidebar link is also sub-optimal, as if you click on it when
you're already on the dashboard, you don't want to reset the query you
have or the "show archived" toggle state - this also requires knowing
the current route.

One surprising thing was that changing the route didn't update the
instanceGroup property of the Dashboard model when I first implemented
it - I recalled I ran into this when implementing the "show archived"
toggle, so it wasn't too bad - but it required explicitly updating the
model in SubPage when the route changed.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>