web: structure: put Route on Session
concourse/concourse#5921

...so that the SideBar has access to the current route. This is useful
to display the active instance group when on the dashboard, and to make
the instance group link smarter by not removing the filter/show archived
toggle.

It also let me clean up the `handleCallback` flow, which was pretty
awkward and required special casing on the subpage in Application in
order to figure out if there's an active pipeline.

I also wanted to apply this refactor to `view`, but I couldn't quite
accomplish that due to OneOffBuilds. Unfortunately, the OneOffBuild
Route does not encode pipeline information, even if the build is a job
build (e.g. /builds/123 actually points to a job build), meaning the
route alone isn't enough to determine the active pipeline in that case.
I wonder if there's a way to transition from the OneOffBuild route to
the Build route after we receive the Build info and determine that it's
a job build (without reloading the page). We *could* also store this on
the SideBar.State, but that seems wrong

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>