web: behaviour: add simple instance group view page
concourse/concourse#5921

It's missing some bells and whistles, but it's a start! Now, clicking on
an instance group filters the dashboard down to the instance groups
pipelines only, which are displayed as individual pipeline cards rather
than an instance group card.

The card headers are still being displayed as the pipeline name (rather
than the instance vars), which looks pretty silly as they're all the
same. Also, we're currently displaying an empty team header that still
takes up space, and a team header in the favorites section, which
both need to be fixed.

This commit also adds the instance group breadcrumb for the instance
group view. It currently does not display the breadcrumb for pipelines
that reside within an instance group, though.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>