introduce NewRouter
I'm not sold on the name, and I find all the ceremony of wrapping a handler in
a router and getting a requestgenerator a bit noisy.

anyway, that gets the direct reference to rata out of this one particular test.

while we're at it, let's get those tests to run a little cheaper by not having
to wait for the OS to allocate/cleanup a socket/listener.

Signed-off-by: Jamie Klassen <jklassen@vmware.com>